Improve unsuccessful/modification of conversation-specific searches
If a conversation-specific search was unsuccessful and/or the user
decided to modify their search, the search form forgot that only a
specific conversation has been searched. This commit changes this by
adding an additional checkbox which remembers the searched conversation
and allows the user to expand the search to all conversations by
unchecking.