Ensure explicit database INDEX names
Checking all revisions of all `.sql` files returned that a INDEX was never
added to an existing table using `ALTER TABLE`. Thus this update script should
not result in changes in *any* installation out there, making this process
quite fast even in large communities. It is added and should be execute
nonetheless for full correctness.

The following fish script was used to check the revisions:

    for file in (g log 2bf40f755228eeae2902562471acf14177777eb6..HEAD --diff-filter=D --no-renames --summary |awk '/delete mode/{print $NF}' |grep '.sql$')
    	for rev in (git rev-list --all -- $file)
    		git show $rev:$file
    	end
    end 2>&1 |grep -v '^fatal:' |grep '^ALTER' |sort -u

see WoltLab/WCF#4505