Fix opening draft conversations
This got broken in 40edd0891d0003c5441dfd2c067eeb5632fbc2ed, because drafts do
not have any participants, thus resulting in an empty list of visible
participants, breaking the database query for the modification log.

Fix this by adding the conversation author to the array if it's a draft
conversation. They are able to see it by definition.